Fix TimeParser and two tests to be able to parse csv like text.
TimeParser : index was not updated after parsing using a Date, so next
check failed and have thrown an error.

In tests, first authors assumed that current locale is US. However used
data won't work with locale like french (a Date has not the same format,
for instance 06/15/1979 in US would be 15/06/1979).
We had to change the code to run it from eclipse, yet also pom.xml to
have the same behavior within maven (programmatic change of locale don't
work with surefire).