#1 Replace existing components with new ones.
As a first step, I tried to find closer widgets in JavaFX from current
Swing implementation.
To have a better result, it could be interesting to compare with a fresh
approach (like FXML, more css styling).

What have to be done also, layouts are not exactly the same. And major
stuff has not be done: translate Visualization and Tree ...