#1 Continue works on renderers.
It becomes difficult to translate Color, and Stroke remains unmanaged.
It might be difficult too to display all shapes we create dynamically as
we should just add them in current Display (Canvas). Yet, it means also
that we have to remove then when needed.

The fact that we have to create a new instance of a Shape every time we
would like to change it could become a good point to redesign the
concept of VisualItem.